pyproject.toml
[tool.poetry]
name = "scissor"
version = "0.1.0"
description = "url shortening"
authors = ["Nana Kwame <stardevone@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.3"
flask-jwt-extended = "^4.5.2"
requests = "^2.31.0"
flask-login = "^0.6.2"
flask-wtf = "^1.1.1"
waitress = "^2.1.2"
install = "^1.3.5"
flask-caching = "^2.0.2"
flask-swagger-ui = "^4.11.1"
pytest = "^7.3.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
